Fix interactive brokers importer
The previous version wrongly assumed that the number of cells per line
is constant across the file. This is only true if the exported file is
first opened and then saved in libreoffice. The raw IB export has a
variable number of cells, and the last field is occasionally not
delimited with a comma.